[TASK] Introduce standalone static TypoScript configuration
Introduced a new standalone static TypoScript configuration
set. Files of of current `default` configuration are moved
into a subfolder, keeping the original files with `@import`
statements to the new location to avoid breaking if project
imported these files directly in a site extension instead of
using (selecting) the static configuration in the page settings.

This make using `EXT:academic_persons` more flexibility for
a standalone usage.