[FEATURE] Add "Card" content element to AcademicPersons extension
Introduced a new "Card" content element to display contacts with
FlexForm configuration.

The `list` flexform is used, but most options disabled as part
of the default PageTsConfig, except the `ProfileList` and the
`fallbackForNonTranslated` options. The last one is important
to have all tools at hand for proper language handling in multi
language instances.

Minor todo's added and missing tests, which needs to be added
with a dedicated change.