[TASK] Streamline `ProfileController` implementation
Code placed in the `initializeAction()` method has only
an effect on the `listAction()` and is completly useless
for the `detailAction()`. `ProfileController` is marked
as `final`, which allows us to move the code within the
overall `initializeAction()` method into the dedicated
method `adoptSetting()` and call it only from the list
action.

This acts a first cleanup to prepare towards having
`storage pages` and `fallbackForNonTranslated` within
the demand object, which is passed to the corresponding
ProfileRepository already, and would remove the need to
create and set the QuerySettings required for specific
list query execetion on a global level and will be done
in a follow up change.