[TASK] Consolidate `ProfileRepository` code into shape
This change consolidates `ProfileRepository` code and
put it into shape as preparation to introduce future
features and changes.

Additionally, demand based flow is adjusted to correctly
emphasize query building when selected profiles are set,
ignoring other filters and sortings in this case, which
was partly in place but not visible on the first look
and matches now how this is handled within the flexform
plugin settings.