[!!!][TASK] Remove project specific `his_id` and `his_parent_id` fields
The `his_id` on a couple of tables is a project
specific addition and not required in general,
and now removed from following tables:

* tx_academicpersons_domain_model_contract
* tx_academicpersons_domain_model_function_type
* tx_academicpersons_domain_model_organisational_unit
* tx_academicpersons_domain_model_location

This includes `TCA` and `ext_tables.sql` configuration
and cleaning the extbase models.

Note that `valid_from` and `valid_to` table column definition
is also removed for `tx_academicpersons_domain_model_location`
in the `ext_tables.sql` file, not having these field defined
in TCA or the extbase model anyway.

This change is technically breaking in case removed his_id and
his_parent_id properties or getter/setter has been used on the
model.