[TASK] Mitigate deprecated `TCEforms` tag in FlexForm
Since TYPO3 v12 the `TCEforms` are deprecated within
FlexForms and should be removed. This is not possible
right now because we need to support TYPO3 v11+v12 in
one version.

To migitate automatic migration and deprecation notices,
flexform are directly provided for both version and only
configured in TCA Overrides based on the current TYPO3
version.

Todo's are added to emphasize what needs to be done when
dropping TYPO3 v11 support.

[1] https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/12.0/Deprecation-97126-TCEformsRemovedInFlexForm.html