[TASK] Replace `DemandInterface` with `ProfileDemand` in `ProfileController`
The `ProfileController` containes different code branches
based on existance of demand object methods not enforced
by the `DemandInterface` but ensured within the concrete
`ProfileDemand` implementation of that interface.

The `ProfileDemand` object is ensured within this final
controller and allows us to use the concrete class as
types and removing conditional based code to simplify
the controller.