[TASK] Use correct API to add cache tags in `ProfileController`
TYPO3 v13 changed the API to set current page cache tags
towards a dedicated request attribute instead of the old
content object renderer.

This change respects this and uses the correct API in the
`ProfileController` and acts as preparation to move forward
to a monorepo and have this code already in place. Does not
have an effect directly, but does not break anything.

[1] https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/13.3/Deprecation-102422-TypoScriptFrontendController-addCacheTags.html