[TASK] Handle deprecated EXT:extbase ViewInterface
The extabse ViewInterface is deprecated and is removed
in TYPO3 v12 and needs proper handling when supporting
v11 and v12 in one version is required like in this
extension.

This change removes native type hints again, replacing
them with proper PHPDoc block annotations and todo's
to mitigate the issue until TYPO3 v11 support is dropped.

[1] https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/11.5/Deprecation-95222-ExtbaseViewInterface.html