[TASK] Set default TCA select item for organisational unit parent
Top organisational units have no parents per design and have
`(int) 0` ensured as `parent` field value when such a record
is created. TCA configuration for `parent` is a foreign table
relation select field and `uid` of zero cannot be retrieved
from the database and due to the missing nature of a existing
record TYPO3 Backend FormEngine honors this by displaying

```
  [Missing label]
```

for this field giving editors the impression of a technical
missconfiguration or error.

To mitigate the missing label this change adjusts the TCA
configuration for the `parent` field by adding a hardcoded
default select item for value `(int) 0` aling with a label
to display the parent state without spreading the impression
something is wrong.