[!!!][TASK] Avoid constructor DI in `AbstractProfileFactory`
Using constructor dependency injection in abstract classes
defines the constructor as API, which should be avoided by
using the inject-method approach and allows to implement
classes using constructor DI without the requirement to deal
and align with parent (abstract) class constructor and passing
it down.

`AbstractProfileFactory` used constructor DI and therefore
violated the above described design pattern.

This change replaces the constructor injection with inject
methods freeing the constructor, which is a breaking change.

The next minor version already contains a couple of breaking
changes and will be communicated as such, which allows us
this change now and mitigate breaking changes related to DI
in the abstract class for the future.

Additionally, the `\Symfony\Contracts\Service\Attribute\Required`
attribute is used for the inject methods to tell symfony DI that
these inject methods needs to be called and are mandatory,
beside having a visually glue for developers. [1]

[1] https://symfony.com/doc/current/service_container/autowiring.html#autowiring-other-methods-e-g-setters-and-public-typed-properties