[TASK] Adjust types in `ModifyListProfilesEvent`
The `ModifyListProfilesEvent` contains a property and
provides a getter for the view object, which has been
based on an extbase related ViewInterface and changed
through multiple TYPO3 and standalone Fluid versions.

This change uses the correct fluid and TYPO3 interfaces
now directly as native types compatible with TYPO3 v12
and v13 and removes the deprecated v11 extbase interface
as a cleanup due to dropped TYPO3 v11 support and solves
the todo comments in the aforementioned class.

That allows us to remove phpstan ignorePatters from the
baselines and further improving the code quality.

In general this follows a similar change made with [1].

Used command(s):

```shell
Build/Scripts/runTests.sh -p 8.1 -t 12 -s composerUpdate \
&& Build/Scripts/runTests.sh -p 8.1 -t 12 -s cgl \
&& Build/Scripts/runTests.sh -p 8.1 -t 12 -s phpstanGenerateBaseline \
&& Build/Scripts/runTests.sh -p 8.2 -t 13 -s composerUpdate \
&& Build/Scripts/runTests.sh -p 8.2 -t 13 -s phpstanGenerateBaseline \
&& Build/Scripts/runTests.sh -p 8.1 -t 12 -s composerUpdate
```

[1] https://github.com/fgtclb/academic-extensions/pull/141