[!!!][FEATURE] Introduce new events for `SelectedProfiles` and `SelectedContracts`
Starting with `2.x` and moving to `TYPO3 v12 & v13` compatibility, two new
plugins has been introduced:

* academicpersons_selectedprofiles
* academicpersons_selectedcontracts

where the `academicpersons_selectedprofiles` was meant as a standalone controllable
plugin in projects as replacedment for the subfeature to select profiles and the
`academicpersons_list` and `academicpersons_listanddetail`, and some code has been
copy & pasted during the implementation of the new actions for the new plugins.

Sadly, dispatching the same `ModifyListProfilesEvent` has also copied over making
it hard in projects or addon extension to distingush for which plugin the event
has been dispatched.

With recently introduced `PluginControllerActionContextInterface` added to the
dispatched event eventlistiner have already easier access to context information,
still making the name of the dispatched event wrong.

This change introduces now dedicated events for the above-mentioned actions, and
removes dispatching the `ModifyListProfilesEvent` for the `selectedProfilesAction`,
which is the `breaking` part of the change.

For the upcoming next minor version an exemption to contain breaking changes have
been already decided making the decision easier. Additionally, we could consider
this with minor impact as most usages will not implementing these events special
for the new one and the risk is considerable low but existing.