[BUGFIX] Ensure to migrate permissions during `list_type` to `CType` migration
TYPO3 v13 deprecated the subtype feature of `tt_content`, also known as
`list_type` for extbase plugins over a long time and will drop it with
TYPO3 v14. To mitigate this, migration in form of an UpgradeWizard has
been implemented, which mystically duplicated itself during the whole
gathering of the academic extensions into the mono-repository.

One of these upgrade wizards contains a hardcoded `return true` for the
`UpgradeWizardInterface::updateNecessary()` method leading to the fact
that it is kept showing up in the available upgrade wizard lists.

The more drastically oversight is, that instances with permissions for
the plugins have not been migrated and left instance admins to verify
and rework the permission for their backend users and groups - a quite
huge effort for bigger instances.

This change strives to mitigate all the aforementioned issues to make
the upgrade experience a breeze in the future, which includes:

* Added new `ListTypeToCTypeUpgradeWizard` based on the upstream TYPO3
  `AbstractListTypeToCTypeUpdate` provided since v13 for all required
  plugins.

* Incomplete and duplicated old upgrade wizards are removed, which are
  `PluginContentUpgrade` and `PluginUpgradeWizard`.

* Functional tests, which has been added for one of the removed upgrade
  wizards are adopted for the new upgrade wizard.

* Functional tests are extended to cover the permission migration.