[TASK] Set `max_line_length = 180` for xlf files in `.editorconfig`
PHPStorm can use the `.editorconfig` set in the mono-repository and
the dedicated academic extension to format files based on the file
type with different settings per file suffix.

This change allows now a maximum line length of 180 characters for
translation files `*.xlf` literally increasing it from the current
default of 80 characters.