[BUGFIX] Use working configuration for `sorting` fields
TYPO3 provides the ability for a dedicated sorting field,
which can be configured using `TCA->ctrl->sortBy` config.
Tables configured with that fields get's automatically a
sorting fields, either directly and also for the relation
and handles this automatically when DataHandler is used.

Even if the field is configured in `ext_tables.sql` and
the extbase domain model has a property `sorting` along
with the getter/setter, reading and persisting the data
does not work out of the box. For a working state it's
mandantory that a column is configured in TCA and due
to the nature it should be `TCAtype=passthrough`.

No table of `EXT:academic_persons` configured `sortBy`
TCA control configuration and most of them even not
the column. Some had a column, but of TCAtype=`none`
which is even worser as that means that there will be
no automatic persistance handling at all.

To mitigate issues due to the missing configuration,
this change:

* Adds `TCA->ctrl->sortBy` configuration for all tables.
* Adds `TCA->columns` column for the sorting field with
  `TCAtype=passthrough`, which includes adopting it for
  existing `TCAtype=none` fields.

No rule without excamption, which means that following
tables are not adjusted because sorting fields does
not make much sense for them:

* `tx_academicpersons_domain_model_function_type`
* `tx_academicpersons_domain_model_location`
* `tx_academicpersons_domain_model_organisational_unit`
* `tx_academicpersons_domain_model_profile`

> [!NOTE]
> The configuration made with this change is elementary
> to have the ability to resort entries using extbase
> domain models, which would not be possible otherwise.