[BUGFIX] ACE-33: Add missing `Repository->findAll()` to repositories
Some repositories already implemented `findAll()` on
their own to ignore storage pages by using following
query setting:

```php
// extbase query
$query = $this->createQuery();
$query->getQuerySettings()->setRespectStoragePage(false);
```

That helps in the plugins to get selectable data, for
example with the frontend edit forms provided by the
`EXT:academic_persons_edit`.

In general, that is not a good solution and we need to
come up with a general working and better approach for
it, but to mitigate some issues within the form handling
of `EXT:academic_persons_edit` we apply the same code
fragment by either adding the `findAll()` method or
adjusting it to the remaining repository implementations.