[TASK] Extract record translation into dedicated `RecordSyncronizationService`
`EXT:academic_persons_edit` implements a PSR-14 event for the
`AfterProfileUpdateEvent` event dispatched in the controller,
with the goal to syncronize profile and relations for languages
configured in the extension configuration.

The internal code within the event listener to handle the sync
for profile and relation records has a couple of issues, which
needs to be investigated, fixed and tested.

Following best practice, event listeners should be glue code
like for example controllers and using services to accomplish
complexer tasks, which is not done here.

This change extracts the handling code from the event listener
of `EXT:academic_persons_edit` into a new services provided by
`EXT:academic_persons` now with a streamlined outer surface,
moving the internal code simply to the new service as private
methods.

Note that the implementation is not touched or changed yet by
intention, which will be done in a dedicated followup along
with added functional test coverage.

The change prepare working on fixing issues and is not itself
a bugfix.