[BUGFIX] Ensure folder/file casing matchies PHP namespace
PHP itself is case-insensitive in reagard of namespace, class,
property, variablen and method/function, which is not ensured
for extended/automatic class autoloading or toolchains around
executing comparison with the excact caseing.

That means, that PHP allows a missmatch in folder and namespace
casing but PHPStan for example eventually not and leading to
class not found issues in projects.

Sadly, recently added DTO interface and default implementation
for controller action context for events in that context shared
the casing `Dto` with other DTO's in the other excentions but
used `DTO` for the folder name leading to the aforementioned
issue in a project not correctly determining the context class
in the dispatched event in `JobController` of `EXT:academic_jobs`.

Naming has not really be decided, but we stay with the `Dto` for
now having the minimal impact and simply align the casing fixing
the static code analyzing issue for projects and avoid a to big
breaking change **without** any benefit and postpone this for the
next major version.