[TASK] ACE-118: Extract logic code from `CreateProfilesCommand` into new service
Some classes, like controllers and commands, are considerable
glue code containers and direct containing logic code should
be reduced to a minimum and instead a suitable service used.

Following that paradigmn this change extracts the logic code
from `CreateProfilesCommand` provided by `academic-persons`
into new dedicated `ProfileCreateCommandService` directly
splitting up the code into internal helper methods keeping
the current logic flow.

`No operation` try-finally blocks are directly added as
preparation to add additional state management for the
environmental state at a later point.

Main benefit of the change is a more encapsulated code
flow, making it easier to add functional tests for it
as the tests can now be done on the service instead of
testing the harder to test glue-code interface of a
command.

On top of that this change also adds a couple of
functional tests for this change and also covering
other things, which has been implemented without
testing because been bound to the command directly.

Further, a couple of bugfixes has been added for
the environment state and build system revealed
by added functional tests.